+++ changelog.md
@@ -1,7 +1,11 @@
 ﻿## 4.3.0
+- Feat: twitter
+	- commands
+	- loop for tweets auto fetch
 - Fix: `commands.md` with wrong links and typos
 - Fix: `messageDelete` deleter not as embed
 - Fix: `date` log field now adapt between prod and test mode
+- Doc : `twitter.md`
 
 ## 4.2.0
 - Feat: add imported gif log detection